pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mosukas_google_drive_autosave"
version = "1.0.0"
authors = [
  { name="Lucas Lima"},
]
description = "A script that automatically saves your projects on google drive after pressing ctrl+s or after a specific time range."
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "altgraph==0.17.4",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "easygui==0.98.3",
    "google-api-core==2.24.0",
    "google-api-python-client==2.155.0",
    "google-auth==2.37.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "googleapis-common-protos==1.66.0",
    "httplib2==0.22.0",
    "idna==3.10",
    "keyboard==0.13.5",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pefile==2023.2.7",
    "proto-plus==1.25.0",
    "protobuf==5.29.1",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pyparsing==3.2.0",
    "pywin32-ctypes==0.2.3",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rsa==4.9",
    "setuptools==75.6.0",
    "six==1.17.0",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
]

[project.optional-dependencies]
creating_exe = [
    "pyinstaller==6.11.1",
    "pyinstaller-hooks-contrib==2024.10",
]

[project.urls]
Homepage = "https://github.com/MosukaSan/Mosukas-Google-Drive-Autosave"
Issues = "https://github.com/MosukaSan/Mosukas-Google-Drive-Autosave/issues"